chore: repurpose info icon
