+++ src/dappsExplorer/DAppsExplorerScreen.tsx
@@ -33,7 +33,7 @@ import { NoResults } from 'src/dappsExplorer/NoResults'
 import useDappFavoritedToast from 'src/dappsExplorer/useDappFavoritedToast'
 import useDappInfoBottomSheet from 'src/dappsExplorer/useDappInfoBottomSheet'
 import useOpenDapp from 'src/dappsExplorer/useOpenDapp'
-import Help from 'src/icons/Help'
+import Help from 'src/icons/navigator/Help'
 import DrawerTopBar from 'src/navigator/DrawerTopBar'
 import { TopBarIconButton } from 'src/navigator/TopBarButton'
 import colors from 'src/styles/colors'
